requirements.txt
jira
python-dotenv
openai
